backend/requirements.txt
# Vulnity-KP Backend Dependencies
# Phase 1: Authentication & Login System

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
# uvloop==0.19.0  # Not supported on Windows

# WebSocket Support
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin bcrypt to compatible version with passlib 1.7.4
python-multipart==0.0.6

# HTTP Client (for future vulnerability scanning)
httpx[http2]==0.25.2

# HTML Parsing (for future vulnerability scanning)
lxml==4.9.3

# Data Validation
pydantic[email]==2.5.0

# Configuration
python-dotenv==1.0.0

# Caching & Rate Limiting (for future use)
redis==5.0.1
aioredis==2.0.1

# Development Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Logging
structlog==23.2.0

# Testing
httpx==0.25.2  # For testing HTTP endpoints
faker==20.1.0  # For generating test data
